Extracted ScoreWindow and ScoreCandidate into OspreySharp.Scoring.CoelutionScorer
* Moved the coelution scorer (ScoreWindow, ScoreCandidate, plus the DIAG_PEPTIDE const and BuildOverridePeaks helper) out of AbstractScoringTask into a new OspreySharp.Scoring.CoelutionScorer; logging and diagnostics are now ctor-injected fields (_logInfo, _diagnostics) instead of threaded parameters
* RunCoelutionScoring constructs the scorer with ctx.LogInfo and the nullable IScoringDiagnostics sink and calls it; AbstractScoringTask shrinks by ~1024 lines
* Behavior-identical relocation; the one substitution is ScoreCandidate's exe-only NUM_PIN_FEATURES (21) -> the Scoring-side OspreyFeatureCalculators.FeatureCount (21)
* Completes the AbstractScoringTask decomposition; gates: 382 + zero-warning; regression Stellar+Astral 1e-9; perf -1.2%; diagnostics-on dump parity identical

See ai/todos/active/TODO-20260611_ospreysharp_decouple_abstractscoring.md
